Updated payment details page to redirect to declaration
