+++ app/views/apply/eligibility/claim/payment-details.html
@@ -18,7 +18,7 @@ Payment details and declaration - {{ serviceName }}
 
   <p>Any money due to you will be paid into your bank account.</p>
 
-  <form action="/apply/{{ claimType }}/eligibility/{{ referenceId }}/claim/{{ claimId }}/payment-details-and-declaration?isAdvance={{ isAdvance }}"
+  <form action="/apply/{{ claimType }}/eligibility/{{ referenceId }}/claim/{{ claimId }}/payment-details?isAdvance={{ isAdvance }}"
         method="post">
 
     <div class="grid-row">
@@ -78,62 +78,12 @@ Payment details and declaration - {{ serviceName }}
       </div>
     </div>
 
-    <h2 class="heading-medium">
-      Declaration
-    </h2>
-
-    {% if isAdvance == 'true' %}
-      <p>I declare that:</p>
-      <ul class="list list-bullet">
-        <li>The information I have given is true to the best of my knowledge</li>
-        <li>The evidence I’ve given is genuine and the most recent version</li>
-        <li>Any money or tickets I receive for a visit will only be used by
-        myself, on the date claimed for</li>
-        <li>If I do not use the tickets or money I will return them to the
-          Assisted Prison Visits Unit (APVU)</li>
-        <li>As this is an advance claim I will submit receipts and a visit
-          confirmation form within 28 days</li>
-        <li>I authorise the APVU to make any necessary checks with the
-          Department for Work and Pensions (DWP)</li>
-        <li>I understand that a false declaration may lead to prosecution and
-          a ban from claiming again</li>
-        <li>I agree to the terms of service as documented in the <a href="/terms-and-conditions" target="_blank">terms and conditions</a></li>
-      </ul>
-    {% else %}
-    <p>I declare that:</p>
-    <ul class="list list-bullet">
-      <li>The information I have given is true to the best of my knowledge </li>
-      <li>The evidence I’ve given is genuine and the most recent version</li>
-      <li>I authorise the APVU to make any necessary checks with the
-      Department for Work and Pensions (DWP)</li>
-      <li>I understand that a false declaration may lead to prosecution and
-      a ban from claiming again</li>
-      <li>I agree to the terms of service as documented in the <a href="/terms-and-conditions" target="_blank">terms and conditions</a></li>
-    </ul>
-    {% endif %}
-
-    <div id="terms-and-conditions" class="form-group{% if errors['terms-and-conditions'][0] %} error {% endif %}">
-      {% if errors['terms-and-conditions'][0] %}
-      <span class="error-message" id="error-message-terms-an-conditions">{{ errors['terms-and-conditions'][0] }}</span>
-      {% endif %}
-      <label for="terms-and-conditions-input" class="block-label">
-        <input id="terms-and-conditions-input"
-               type="checkbox"
-               name="terms-and-conditions-input"
-               {% if bankDetails['tAndC'] %}
-                checked
-               {% endif %}
-               value="yes">
-        <span class="heading-small"> I agree to the declaration</span>
-      </label>
-    </div>
-
     {% include "partials/csrf-hidden-input.html" %}
 
     <input id="claim-submit"
            type="submit"
            class="button"
-           value="Finish">
+           value="Continue">
   </form>
 
 </main>
