+++ test/unit/routes/apply/eligibility/claim/test-claim-summary.js
@@ -86,12 +86,12 @@ describe('routes/apply/eligibility/claim/claim-summary', function () {
         })
     })
 
-    it('should respond with a 302 to payment details and declaration', function () {
+    it('should respond with a 302 to payment details', function () {
       getClaimSummaryStub.resolves(CLAIM)
       return supertest(app)
         .post(ROUTE)
         .expect(302)
-        .expect('location', `/apply/${CLAIM_TYPE}/eligibility/${REFERENCE_ID}/claim/${CLAIM_ID}/payment-details-and-declaration?isAdvance=false`)
+        .expect('location', `/apply/${CLAIM_TYPE}/eligibility/${REFERENCE_ID}/claim/${CLAIM_ID}/payment-details?isAdvance=false`)
     })
 
     it('should respond with a 400 if validation errors', function () {
